feat(pipe): allow the output of the fetch script to be piped into the solver
Adds the -p and -f options to the fetch script, which allows the results
to be printed to stdout or to a file. The solver script has been
adjusted to be able to read from stdin as well. This allows for
something like: python fetch_data.py -p | python dfs_solver.py